Corretti nomi
Bisogna sistemare lo static in alcuni metodi
