--- src/main/java/it/SWEasabi/authentication/kernel/Blacklister.java
@@ -1,15 +1,0 @@
-package it.SWEasabi.authentication.kernel;
-
-import it.SWEasabi.authentication.services.BlackListService;
-
-public class Blacklister 
-{
-    public static void blacklist(BlackListService service, String refreshJwt)
-    {
-        service.blacklist(refreshJwt);
-    }
-    public static boolean isBlacklisted(BlackListService service, String refreshJwt)
-    {
-        return service.isBlacklisted(refreshJwt);
-    }
-}
