+++ src/main/java/it/SWEasabi/authentication/kernel/JwtExtractor.java
@@ -9,21 +9,21 @@ import com.auth0.jwt.interfaces.DecodedJWT;
 
 public class JwtExtractor
 {
-    private static Map<String, Object> getClaimsMap(String token, String signature) throws Exception
+    private static Map<String, Object> getClaimsMap(String jwt, String key) throws Exception
     {
-        JWTVerifier verifier = JWT.require(Algorithm.HMAC256(signature))
+        JWTVerifier verifier = JWT.require(Algorithm.HMAC256(key))
                 .withIssuer("auth0")
                 .withClaimPresence("map")
                 .withClaimPresence("type")
                 .build();
-        DecodedJWT jwt = verifier.verify(token);
-        Map<String, Object> map = jwt.getClaim("map").asMap();
+        DecodedJWT decodedJWT = verifier.verify(jwt);
+        Map<String, Object> map = decodedJWT.getClaim("map").asMap();
         
         return map;
     }
-    public static String getUsername(String token, String signature) throws Exception
+    public static String getUsername(String jwt, String key) throws Exception
     {
-        Map<String, Object> map = getClaimsMap(token, signature);
+        Map<String, Object> map = getClaimsMap(jwt, key);
         return map.get("username").toString();
     }
 }
