+++ src/main/java/it/SWEasabi/authentication/kernel/JwtVerifier.java
@@ -10,31 +10,31 @@ import com.auth0.jwt.exceptions.SignatureVerificationException;
 import com.auth0.jwt.exceptions.MissingClaimException;
 */
 
-import it.SWEasabi.authentication.services.BlackListService;
+import it.SWEasabi.authentication.services.BlacklistService;
 import it.SWEasabi.authentication.services.KeysService;
 
 public class JwtVerifier
 {
     private KeysService keys;
-    private BlackListService blackListService;
-    public JwtVerifier(KeysService Keys, BlackListService BlackListService)
+    private BlacklistService blacklistService;
+    public JwtVerifier(KeysService Keys, BlacklistService BlacklistService)
     {
         keys = Keys;
-        blackListService = BlackListService;
+        blacklistService = BlacklistService;
     }
 
-    private boolean verify(String token, String signature, String type)
+    private boolean verify(String jwt, String key, String type)
     {
         try
         {
-            JWTVerifier verifier = JWT.require(Algorithm.HMAC256(signature))
+            JWTVerifier verifier = JWT.require(Algorithm.HMAC256(key))
                 .withIssuer("auth0")
                 .withClaimPresence("map")
                 .withClaimPresence("type")
                 .build();
-            DecodedJWT jwt = verifier.verify(token);
+            DecodedJWT decodedJWT = verifier.verify(jwt);
 
-            if(jwt.getClaim("type").asString().equals(type))
+            if(decodedJWT.getClaim("type").asString().equals(type))
             {
                 return true;
             }
@@ -45,16 +45,16 @@ public class JwtVerifier
         }
         return false;
     }
-    public boolean isAValidAccessJwt(String token)
+    public boolean isAValidAccessJwt(String jwt)
     {
-        return verify(token, keys.getAccessKey(), "access");
+        return verify(jwt, keys.getAccessKey(), "access");
     }
-    public boolean isAValidRefreshJwt(String token)
+    public boolean isAValidRefreshJwt(String jwt)
     {
-        if(Blacklister.isBlacklisted(blackListService, token))
+        if(RefreshJwtBlacklister.isBlacklisted(blacklistService, jwt))
         {
             return false;
         }
-        return verify(token, keys.getRefreshKey(), "refresh");
+        return verify(jwt, keys.getRefreshKey(), "refresh");
     }
 }
