+++ src/main/java/it/SWEasabi/authentication/kernel/RefreshJwtBlacklister.java
@@ -1,0 +1,15 @@
+package it.SWEasabi.authentication.kernel;
+
+import it.SWEasabi.authentication.services.BlacklistService;
+
+public class RefreshJwtBlacklister 
+{
+    public static void blacklist(BlacklistService service, String refreshJwt)
+    {
+        service.blacklist(refreshJwt);
+    }
+    public static boolean isBlacklisted(BlacklistService service, String refreshJwt)
+    {
+        return service.isBlacklisted(refreshJwt);
+    }
+}
